fix: move constants folder && remove unnecessary variables
