--- src/features/constants/common.ts
@@ -1,182 +1,0 @@
-export const pastelTableNames = {
-  statisticinfo: 'statisticinfo',
-  networkinfo: 'networkinfo',
-  nettotals: 'nettotals',
-  mempoolinfo: 'mempoolinfo',
-  rawmempoolinfo: 'rawmempoolinfo',
-  mininginfo: 'mininginfo',
-  blockinfo: 'blockinfo',
-  rawtransaction: 'rawtransaction',
-  transaction_tbl: 'transaction_tbl',
-  txoutsetinfo: 'txoutsetinfo',
-  chaintips: 'chaintips',
-  blocksubsidy: 'blocksubsidy',
-  walletinfo: 'walletinfo',
-  listreceivedbyaddress: 'listreceivedbyaddress',
-  listtransactions: 'listtransactions',
-  listunspent: 'listunspent',
-  totalbalance: 'totalbalance',
-  listaddresses: 'listaddresses',
-}
-
-export const statisticsFieldNames = [
-  { name: 'Mempoolinfo', routeName: '/mempoolinfo' },
-  { name: 'Hashrate', routeName: '/hashrateovertime' },
-  { name: 'Blockinfo', routeName: '/blockinfo' },
-  { name: 'Rawtransaction', routeName: '/rawtransaction' },
-  { name: 'Txoutsetinfo', routeName: '/txoutsetinfo' },
-  { name: 'Walletinfo', routeName: '/walletinfo' },
-]
-
-export const mempoolinfoSizeRange = [
-  { key: 'y1', value: 1 },
-  { key: 'y2', value: 2 },
-  { key: 'y3', value: 3 },
-  { key: 'y4', value: 4 },
-  { key: 'y5', value: 5 },
-  { key: 'y6', value: 6 },
-  { key: 'y7', value: 7 },
-  { key: 'y8', value: 8 },
-  { key: 'y10', value: 10 },
-  { key: 'y12', value: 12 },
-  { key: 'y15', value: 15 },
-  { key: 'y20', value: 20 },
-  { key: 'y30', value: 30 },
-  { key: 'y40', value: 40 },
-  { key: 'y50', value: 50 },
-  { key: 'y60', value: 60 },
-  { key: 'y70', value: 70 },
-  { key: 'y80', value: 80 },
-  { key: 'y100', value: 100 },
-  { key: 'y125', value: 125 },
-  { key: 'y150', value: 150 },
-  { key: 'y175', value: 175 },
-  { key: 'y200', value: 200 },
-  { key: 'y250', value: 250 },
-  { key: 'y300', value: 300 },
-  { key: 'y400', value: 400 },
-  { key: 'y500', value: 500 },
-  { key: 'y600', value: 600 },
-  { key: 'y700', value: 700 },
-  { key: 'y800', value: 800 },
-  { key: 'y1000', value: 1000 },
-  { key: 'y1200', value: 1200 },
-  { key: 'y1400', value: 1400 },
-  { key: 'y1700', value: 1700 },
-  { key: 'y2000', value: 2000 },
-]
-
-export const mempoolAreainfoKeys = [
-  'y1',
-  'y2',
-  'y3',
-  'y4',
-  'y5',
-  'y6',
-  'y7',
-  'y8',
-  'y10',
-  'y12',
-  'y15',
-  'y20',
-  'y30',
-  'y40',
-  'y50',
-  'y60',
-  'y70',
-  'y80',
-  'y100',
-  'y125',
-  'y150',
-  'y175',
-  'y200',
-  'y250',
-  'y300',
-  'y400',
-  'y500',
-  'y600',
-  'y700',
-  'y800',
-  'y1000',
-  'y1200',
-  'y1400',
-  'y1700',
-  'y2000',
-]
-
-export const mempoolinfoColors = {
-  y1: '#0000EC',
-  y2: '#0000EC',
-  y3: '#0000EC',
-  y4: '#0000EC',
-  y5: '#0000FF',
-  y6: '#2C2CFF',
-  y7: '#5858FF',
-  y8: '#8080FF',
-  y10: '#008000',
-  y12: '#00A000',
-  y15: '#00E000',
-  y20: '#30E030',
-  y30: '#90E090',
-  y40: '#808000',
-  y50: '#989800',
-  y60: '#B0B000',
-  y70: '#C8C800',
-  y80: '#E0E000',
-  y100: '#E0E030',
-  y125: '#E0E060',
-  y150: '#800000',
-  y175: '#A00000',
-  y200: '#C00000',
-  y250: '#E00000',
-  y300: '#E02020',
-  y400: '#E04040',
-  y500: '#E06060',
-  y600: '#800080',
-  y700: '#AC00AC',
-  y800: '#D800D8',
-  y1000: '#FF00FF',
-  y1200: '#FF2CFF',
-  y1400: '#FF58FF',
-  y1700: '#FF80FF',
-  y2000: '#000000',
-}
-
-export const transformInit = {
-  time: null,
-  y1: 0,
-  y2: 0,
-  y3: 0,
-  y4: 0,
-  y5: 0,
-  y6: 0,
-  y7: 0,
-  y8: 0,
-  y10: 0,
-  y12: 0,
-  y15: 0,
-  y20: 0,
-  y30: 0,
-  y40: 0,
-  y50: 0,
-  y60: 0,
-  y70: 0,
-  y80: 0,
-  y100: 0,
-  y125: 0,
-  y150: 0,
-  y175: 0,
-  y200: 0,
-  y250: 0,
-  y300: 0,
-  y400: 0,
-  y500: 0,
-  y600: 0,
-  y700: 0,
-  y800: 0,
-  y1000: 0,
-  y1200: 0,
-  y1400: 0,
-  y1700: 0,
-  y2000: 0,
-}
