+++ src/features/pastelDB/constants/common.ts
@@ -1,0 +1,20 @@
+export const pastelTableNames = {
+  statisticinfo: 'statisticinfo',
+  networkinfo: 'networkinfo',
+  nettotals: 'nettotals',
+  mempoolinfo: 'mempoolinfo',
+  rawmempoolinfo: 'rawmempoolinfo',
+  mininginfo: 'mininginfo',
+  blockinfo: 'blockinfo',
+  rawtransaction: 'rawtransaction',
+  transaction_tbl: 'transaction_tbl',
+  txoutsetinfo: 'txoutsetinfo',
+  chaintips: 'chaintips',
+  blocksubsidy: 'blocksubsidy',
+  walletinfo: 'walletinfo',
+  listreceivedbyaddress: 'listreceivedbyaddress',
+  listtransactions: 'listtransactions',
+  listunspent: 'listunspent',
+  totalbalance: 'totalbalance',
+  listaddresses: 'listaddresses',
+}
